refactor: applied user prompting in sc2egset_replaypack_processor
