+++ src/datasetpreparator/sc2/sc2egset_replaypack_processor/sc2egset_replaypack_processor.py
@@ -1,17 +1,18 @@
-import os
 from pathlib import Path
 import logging
 import click
 from tqdm import tqdm
 
+from datasetpreparator.sc2.sc2egset_replaypack_processor.utils.download_maps import (
+    sc2info_extractor_go_map_download,
+)
 from datasetpreparator.settings import LOGGING_FORMAT
 from datasetpreparator.sc2.sc2egset_replaypack_processor.utils.replaypack_processor_args import (
     ReplaypackProcessorArguments,
-    SC2InfoExtractorGoArguments,
+    define_sc2egset_args,
 )
 from datasetpreparator.sc2.sc2egset_replaypack_processor.utils.multiprocess import (
     multiprocessing_scheduler,
-    pre_process_download_maps,
 )
 
 from datasetpreparator.directory_flattener.directory_flattener import (
@@ -27,57 +28,7 @@ from datasetpreparator.processed_mapping_copier.processed_mapping_copier import 
 )
 
 from datasetpreparator.file_renamer.file_renamer import file_renamer
-
-
-def define_sc2egset_args(
-    input_path: Path,
-    output_path: Path,
-    arguments: ReplaypackProcessorArguments,
-    maybe_dir: Path,
-) -> ReplaypackProcessorArguments | None:
-    logging.debug(f"Processing entry: {maybe_dir}")
-    processing_input_dir = Path(input_path, maybe_dir).resolve()
-    if not processing_input_dir.is_dir():
-        logging.debug("Entry is not a directory, skipping!")
-        return None
-
-    logging.debug(f"Output dir: {output_path}")
-    # Create the main output directory:
-    if not output_path.exists():
-        output_path.mkdir()
-
-    # TODO: use pathlib:
-    path, output_directory_name = os.path.split(maybe_dir)
-    logging.debug(f"Output dir name: {output_directory_name}")
-    if output_directory_name == "input":
-        return None
-
-    output_directory_with_name = Path(output_path, output_directory_name).resolve()
-    logging.debug(f"Output filepath: {output_directory_with_name}")
-
-    # Create the output subdirectories:
-    if not output_directory_with_name.exists():
-        output_directory_with_name.mkdir()
-
-    sc2_info_extractor_go_args = (
-        SC2InfoExtractorGoArguments.get_sc2egset_processing_args(
-            processing_input=processing_input_dir,
-            output=output_directory_with_name,
-            perform_chat_anonymization=arguments.perform_chat_anonymization,
-        )
-    )
-
-    return sc2_info_extractor_go_args
-
-
-def sc2info_extractor_go_map_download(arguments: ReplaypackProcessorArguments):
-    # Pre-process, download all maps:
-    logging.info("Downloading all maps...")
-    map_download_arguments = SC2InfoExtractorGoArguments.get_download_maps_args(
-        processing_input=arguments.input_path, output=arguments.output_path
-    )
-    pre_process_download_maps(arguments=map_download_arguments)
-    pass
+from datasetpreparator.utils.user_prompt import user_prompt_overwrite_ok
 
 
 def sc2egset_replaypack_processor(
@@ -93,15 +44,9 @@ def sc2egset_replaypack_processor(
         Specifies the arguments as per the ReplaypackProcessorArguments class fields.
     """
 
-    input_path = arguments.input_path
-    output_path = arguments.output_path
-    n_processes = arguments.n_processes
-
     multiprocessing_list = []
-    for maybe_dir in tqdm(list(input_path.iterdir())):
+    for maybe_dir in tqdm(list(arguments.input_path.iterdir())):
         sc2_info_extractor_go_args = define_sc2egset_args(
-            input_path=input_path,
-            output_path=output_path,
             arguments=arguments,
             maybe_dir=maybe_dir,
         )
@@ -109,7 +54,7 @@ def sc2egset_replaypack_processor(
             multiprocessing_list.append(sc2_info_extractor_go_args)
 
     # Run processing with multiple SC2InfoExtractorGo instances:
-    multiprocessing_scheduler(multiprocessing_list, int(n_processes))
+    multiprocessing_scheduler(multiprocessing_list, int(arguments.n_processes))
 
 
 @click.command(
@@ -147,6 +92,13 @@ def sc2egset_replaypack_processor(
     help="Number of processes to be spawned for the dataset processing with SC2InfoExtractorGo.",
 )
 @click.option(
+    "--force_overwrite",
+    type=bool,
+    default=False,
+    required=True,
+    help="Flag that specifies if the user wants to overwrite files or directories without being prompted.",
+)
+@click.option(
     "--log",
     type=click.Choice(["INFO", "DEBUG", "ERROR", "WARN"], case_sensitive=False),
     default="WARN",
@@ -156,6 +108,7 @@ def main(
     input_path: Path,
     output_path: Path,
     n_processes: int,
+    force_overwrite: bool,
     log: str,
 ) -> None:
     numeric_level = getattr(logging, log.upper(), None)
@@ -164,7 +117,7 @@ def main(
     logging.basicConfig(format=LOGGING_FORMAT, level=numeric_level)
 
     # Create output directory if it does not exist:
-    if not output_path.exists():
+    if user_prompt_overwrite_ok(filepath=output_path, force=force_overwrite):
         output_path.mkdir(exist_ok=True)
 
     input_path = Path(input_path).resolve()
@@ -193,7 +146,9 @@ def main(
     # input command waiting for user input to confirm potential overwrite:
     #
 
-    if not directory_flattener_output_path.exists():
+    if user_prompt_overwrite_ok(
+        filepath=directory_flattener_output_path, force=force_overwrite
+    ):
         directory_flattener_output_path.mkdir(exist_ok=True)
 
     # TODO: Check if the output directory is not empty, if it is you can proceed
