+++ src/datasetpreparator/sc2/sc2egset_replaypack_processor/utils/replaypack_processor_args.py
@@ -1,3 +1,4 @@
+import logging
 from pathlib import Path
 import os
 
@@ -170,3 +171,63 @@ class ReplaypackProcessorArguments:
         self.output_path = output_path
         self.n_processes = n_processes
         self.maps_directory = maps_directory
+
+
+def define_sc2egset_args(
+    arguments: ReplaypackProcessorArguments,
+    maybe_dir: Path,
+) -> ReplaypackProcessorArguments | None:
+    """
+    Creates final ReplaypackProcessorArguments for SC2InfoExtractorGo. These arguments
+    are used to perform extraction of the data from a replaypack with the same arguments
+    as SC2EGSetDataset.
+
+    Parameters
+    ----------
+    arguments : ReplaypackProcessorArguments
+        Arguments to the command line tool
+    maybe_dir : Path
+        _description_
+
+    Returns
+    -------
+    ReplaypackProcessorArguments | None
+        _description_
+    """
+
+    input_path = arguments.input_path
+    output_path = arguments.output_path
+
+    logging.debug(f"Processing entry: {maybe_dir}")
+    processing_input_dir = Path(input_path, maybe_dir).resolve()
+    if not processing_input_dir.is_dir():
+        logging.debug("Entry is not a directory, skipping!")
+        return None
+
+    logging.debug(f"Output dir: {output_path}")
+    # Create the main output directory:
+    if not output_path.exists():
+        output_path.mkdir()
+
+    # TODO: use pathlib:
+    path, output_directory_name = os.path.split(maybe_dir)
+    logging.debug(f"Output dir name: {output_directory_name}")
+    if output_directory_name == "input":
+        return None
+
+    output_directory_with_name = Path(output_path, output_directory_name).resolve()
+    logging.debug(f"Output filepath: {output_directory_with_name}")
+
+    # Create the output subdirectories:
+    if not output_directory_with_name.exists():
+        output_directory_with_name.mkdir()
+
+    sc2_info_extractor_go_args = (
+        SC2InfoExtractorGoArguments.get_sc2egset_processing_args(
+            processing_input=processing_input_dir,
+            output=output_directory_with_name,
+            perform_chat_anonymization=arguments.perform_chat_anonymization,
+        )
+    )
+
+    return sc2_info_extractor_go_args
